requirements.txt
crewai==1.14.1
openai==2.32.0
python-dotenv==1.1.1
boto3==1.42.90
requests==2.32.5
